+++ src/pages/LogView/LoadingPage/useResolveLogURLAndRenderingType.ts
